% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_cpi.R
\name{eb_cpi}
\alias{eb_cpi}
\title{Expenditure-Based Index}
\usage{
eb_cpi(price_data, expenditure_data)
}
\arguments{
\item{price_data}{Dataset of prices that must contain columns of a base and current year prices.
There must be equal number of base and current year price variables.}

\item{expenditure_data}{Dataset containing List of items and their respective
expenditure prices. Consequently a price dataset with only two columns/variables;
\code{item} and \code{expenditure}. This dataset is the base year expenditure prices from \code{price_data}.
Hence the number of columns for the base year prices in \code{price_data} should correspond to
the number of rows in \code{expenditure_data}.}
}
\value{
a float number representing the Laspeyres' index
}
\description{
Expenditure-Based Index
}
\examples{
eb_cpi(price_data = empirical_Data1, expenditure_data = expenditure_Data)
eb_cpi(empirical_Data1, expenditure_Data)
}
