% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_Data2.R
\docType{data}
\name{empirical_Data2}
\alias{empirical_Data2}
\title{Pseudo Price Data}
\format{
\subsection{\code{empirical_Data2}}{

A data frame with 89 rows and 110 columns:
\describe{
\item{B1,B2,B3...B55}{a simulated price data of 55 different items. the columns B1 to B55
denotes price data for a base year of 2002.}
\item{C1,C2,C3...C55}{a simulated price data of 55 different items. the columns C1 to C55
denotes price data for a current year of 2012.}
...
}
}
}
\source{
\url{https://www.numbeo.com/cost-of-living/}
}
\usage{
empirical_Data2
}
\description{
A mock price data of 55 base and current year prices simulated from using
mean vector and covariance matrix of price data
obtained from the internet ...
}
\keyword{datasets}
