% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenditure_Data.R
\docType{data}
\name{expenditure_Data}
\alias{expenditure_Data}
\title{Expenditure Price Data}
\format{
A data frame with 19 rows and 2 columns:
\describe{
\item{Item}{These are the 19 randomly selected items from the basket of the
Ghana Statistical Services. Observations P1,P2...P19 corresponds to the same 19 items
found in \code{empirical_Data1}}

\item{Expenditure}{These are price values captured under the base year prices for
\code{empirical_Data1}}
...
}
}
\source{
{Obtained from the Ghana Statistical Services head office in Accra}
}
\usage{
expenditure_Data
}
\description{
This is the price data of the 19 items from \code{empirical_Data1}. It contains only
the base year prices from \code{empirical_Data1}.
Ideally, The expenditure data is collected separately by survey. It's collected once in
every 5 to 10 years. This dataset however, is a self manufactured expenditure dataset.
}
\keyword{datasets}
