% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvw_cpi.R
\name{mvw_cpi}
\alias{mvw_cpi}
\title{Multivariate Weighted Indices}
\usage{
mvw_cpi(data = random_data)
}
\arguments{
\item{data}{Dataset of prices that must contain columns for
a base year and current year prices. There must be equal number
of base and current year price variables.}
}
\value{
a vector of indices for calculating consumer price index (CPI)
}
\description{
Multivariate Weighted Indices
}
\examples{
mw_indices()
mw_indices(data = empirical_Data1)
mw_indices(data = empirical_Data2)
}
