% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvw_inflation.R
\name{mvw_inflation}
\alias{mvw_inflation}
\title{Inflation Value from multivariate weighted indices}
\usage{
mvw_inflation(index = c("fisher", "paashe", "laspeyres", "drobish"), data)
}
\arguments{
\item{index}{takes any of the indexes from the multivariate weighted indices that is:
the output from the \code{mvw_cpi} function, which are either Fishers index (fisher),
Paashe's index (paashe),Laspeyres index (laspeyres) or Drobish Bowleys index (drobish)}

\item{data}{price dataset provided by user from which inflation is calculated. Must be in
the same format as that used in the \code{mvw_cpi} function.}
}
\value{
the inflation value for the period used to determine mvw_cpi indices
}
\description{
Inflation Value from multivariate weighted indices
}
\examples{
# generates a warning message if index provided is not fishers'
mvw_inflation('drobish', empirical_Data1)
mvw_inflation(index = 'fisher', data = empirical_Data1)
}
